requirements.txt
torch==2.0.1
requests
pillow